Fix compilation on OS X / Homebrew clang
The default Homebrew installation of `clang` on OS X recommends setting
`CXXFLAGS` to ensure that the correct standard library implementation is
used (`libc++` instead of `stdlibc++`). However, these flags are
incompatible with `g++`, which breaks this build.

To fix this, avoid explicitly setting the compiler executable (which at
any rate should not be necessary).